refactor: rename extract to utils
