+++ packages/rust-runtime/src/lib.rs
@@ -2,4 +2,4 @@ pub mod http;
 pub mod result;
 pub mod runtime;
 
-mod extract;
+mod utils;
