+++ packages/rust-runtime/src/utils.rs
